+++ src/main/java/com/ripple/BE/news/application/crawler/impl/InvestmentNewsCrawler.java
@@ -1,6 +1,6 @@
-package com.ripple.BE.news.crawler.impl;
+package com.ripple.BE.news.application.crawler.impl;
 
-import com.ripple.BE.news.crawler.NewsCrawler;
+import com.ripple.BE.news.application.crawler.NewsCrawler;
 import com.ripple.BE.news.domain.type.NewsCategory;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -9,7 +9,8 @@ import org.springframework.stereotype.Component;
 @Component
 public class InvestmentNewsCrawler extends NewsCrawler {
 
-    protected static final String URL = "https://news.naver.com/breakingnews/section/101/258";
+    protected static final String URL =
+            "https://news.naver.com/main/list.naver?mode=LS2D&mid=sec&sid1=101&sid2=258&listType=paper";
 
     /**
      * 페이지 URL
